Relocate strings
