+++ Sources/ArcGISToolkit/Components/FeatureFormView/FeatureFormView.swift
@@ -468,71 +468,3 @@ extension Logger {
         Logger(subsystem: "com.esri.ArcGISToolkit", category: "FeatureFormView")
     }
 }
-
-extension Text {
-    static var exceedsMaximumAttachmentDurationMessage: Self? {
-        let maxDuration = Duration.seconds(1)
-        return .init(
-            "The media is too long. Maximum allowed duration is ^[\(maxDuration.components.seconds) seconds](inflect: true).",
-            bundle: .toolkitModule,
-            comment: "A message indicating that no more attachments can be added because the maximum count has been reached."
-        )
-    }
-    
-    static var incorrectAttachmentTypeMessage: Self {
-        .init(
-            "This file type is not supported.",
-            bundle: .toolkitModule,
-            comment: "A message indicating the type of the supplied attachment is not supported.."
-        )
-    }
-    
-    func maxAttachmentCountMessage(max: Int) -> Self {
-        return .init(
-            "The maximum number of attachments allowed is \(max).",
-            bundle: .toolkitModule,
-            comment: "A message indicating that no more attachments can be added because the maximum count has been reached."
-        )
-    }
-    
-    static var maxFileSizeMessage: Self? {
-        let maxFileSize: Measurement<UnitInformationStorage> = .init(value: 1, unit: .megabytes)
-        return .init(
-            "The file is too large. Maximum allowed file size is \(maxFileSize) MB.",
-            bundle: .toolkitModule,
-            comment: "A message indicating that no more attachments can be added because the maximum count has been reached."
-        )
-    }
-    
-    func minAttachmentCountMessage(min: Int) -> Self {
-        if min == 1 {
-            return .init(
-                "At least \(min) attachment is required.",
-                bundle: .toolkitModule,
-                comment: "A message indicating the element requires at least one attachment."
-            )
-        } else {
-            return .init(
-                "At least \(min) attachments are required.",
-                bundle: .toolkitModule,
-                comment: "A message indicating the element requires multiple attachments."
-            )
-        }
-    }
-    
-    static var takePhoto: Self {
-        .init(
-            "Take Photo",
-            bundle: .toolkitModule,
-            comment: "A label for a button to capture a new photo."
-        )
-    }
-    
-    static var takeVideo: Self {
-        .init(
-            "Take Video",
-            bundle: .toolkitModule,
-            comment: "A label for a button to capture a new video."
-        )
-    }
-}
